requirements.txt
elevenlabs
openai
moviepy
pexels-api-py
faster-whisper
opencv-python
mutagen
